zxtune-qt: build fix.
git-svn-id: https://zxtune.googlecode.com/svn/trunk@2738 e19a5227-2556-0410-816f-a35683127991

+++ apps/zxtune-qt/playlist/ui/desktop/playlist_contextmenu.cpp
@@ -118,7 +118,7 @@ namespace
     }
     else
     {
-      return QFileSystemModelTranslate("%1 bytes").arg(QLocale().toString(size));
+      return QFileSystemModelTranslate("%1 bytes").arg(QLocale().toString(qulonglong(size)));
     }
   }
